implemented push and pop for minheap structure
So far I have updated the push function to add the idx parameter(value of frequency of char) to the end of the array which is next "open" spot of the minheap structure. Also, the pop structure has been updated to replace the root of the minheap with the last element of the min heap. Size was also increased and decresed for both function.